moved msal to lib/ since its a service not api request
